refactor new Player
Extract some common audiocontext features to an AudioContext class.
Player extends this class. This makes it more convenient to refactor/add
a simple recorder later.
