+++ src/WebAudio/Player.js
@@ -1,17 +1,11 @@
 import {authorisedRequest, request} from '../api/communication';
-import debug from 'debug';
+import AudioContext from './AudioContext';
 
 /**
  * Player
- * Audio player based on WebAudio technology
+ * Simple Audio Player based on Web Audio API technology
  */
-export default class Player {
-  /**
-   * Private object to hold AudioContext node.
-   * @private
-   */
-  audioContext = null;
-
+export default class Player extends AudioContext {
   /**
    * Private object to hold AudioBuffer node.
    * @private
@@ -23,143 +17,6 @@ export default class Player {
    * @private
    */
   audioSource = null;
-
-  /**
-   * Keep a list of all registered event listeners.
-   * This way, we could conveniently remove all the listeners at once.
-   * @type {Array}
-   * @private
-   */
-  eventListeners = [];
-
-  /**
-   * Set logging namespace.
-   * @private
-   */
-  log = debug('its-sdk:AudioPlayer');
-
-  /**
-   * Construct the Player.
-   *
-   * @param {Object} audioContext - Allow to provide custom/own audioContext object.
-   */
-  constructor(audioContext = null) {
-    // Provide AudioContext object. If not passed the
-    // default window.AudioContext will be used.
-    if (audioContext) {
-      this.audioContext = audioContext;
-    } else {
-      this.audioContext = this.createAudioContext();
-    }
-
-    this.suspendAudioContext = this.suspendAudioContext.bind(this);
-
-    // set log to output to stdout
-    this.log.log = console.log.bind(console);
-  }
-
-  /**
-   * Wraps the addEventListener which is available on the AudioContext node.
-   *
-   * @see https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/attachEvent
-   * @param {...*} args - Array with passed arguments.
-   */
-  addEventListener(...args) {
-    // First, add the event listener to our local list.
-    this.eventListeners.push(args);
-
-    // Now, actually add the event listener!
-    this.audioContext.addEventListener(...args);
-  }
-
-  /**
-   * Wraps the removeEventListener which is available on the AudioContext node.
-   * Make sure to call with the same arguments as the addEventListner.
-   *
-   * @see https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/removeEventListener
-   * @param {...*} args - Array with passed arguments.
-   */
-  removeEventListener(...args) {
-    // First, try to find the event in our list.
-    const itemIndex = this.eventListeners.findIndex(item => {
-      const items = item.length;
-      let count = 0;
-
-      // If the count of items doesn't equal, it's definitely not the
-      // listener we want to remove.
-      if (items !== args.length) {
-        return false;
-      }
-
-      // Start looping through all items.
-      for (count; count <= items; count += 1) {
-        if (typeof item[count] === 'function') {
-          if (typeof args[count] !== 'function') {
-            // typecheck failed, so return false!
-            return false;
-          } else if (item[count].toString() !== args[count].toString()) {
-            // Both are function, but not the same!
-            return false;
-          }
-        } else if (item[count] !== args[count]) {
-          // values do not equal
-          return false;
-        }
-      }
-
-      // If we hit here, we found a match! On fire!
-      return true;
-    });
-
-    // Remove that item from the list
-    if (itemIndex) {
-      this.eventListeners.splice(itemIndex, 1);
-    }
-
-    // Remove the event handler!
-    this.audioContext.removeEventListener(...args);
-  }
-
-  removeAllEventListeners() {
-    this.eventListeners.forEach(listener => {
-      this.removeEventListener(...listener);
-    });
-    this.eventListeners = [];
-  }
-
-  /**
-   * Use this method to conveniently fire an event.
-   * We could, if we wanted, add some data.
-   *
-   * @param {string} eventName - Event to fire.
-   * @param {Object} data - Data to pass as detail.
-   * @private
-   */
-  fireEvent(eventName = null, data = null) {
-    if (!eventName) {
-      return;
-    }
-
-    this.audioContext.dispatchEvent(new CustomEvent(eventName, {detail: data}));
-  }
-
-  /**
-   * Get the audio context or create one.
-   *
-   * @return {AudioContext} The AudioContext created will be returned
-   */
-  createAudioContext() {
-    if (!window.ItslAudioContext) {
-      window.AudioContext =
-        window.AudioContext || window.webkitAudioContext;
-      window.ItslAudioContext = new window.AudioContext();
-    }
-    return window.ItslAudioContext;
-  }
-
-  suspendAudioContext() {
-    this.audioContext.suspend();
-  }
 
   createBufferSource() {
     this.disconnectBufferSource();
@@ -213,7 +70,7 @@ export default class Player {
         this.fireEvent('loaded');
       });
     } catch (error) {
-      this.log(`${error.name}: ${error.message}`);
+      this.error(`${error.name}: ${error.message}`);
     }
   }
